# Speed up build by not building examples and tests