Building the default target which includes examples and tests:
```
$ time make
[...]
make  55.96s user 6.75s system 91% cpu 1:08.67 total
```

Only using the build target:
```
$ time make build
[...]
make build  37.11s user 2.96s system 98% cpu 40.699 total
```