@pyramation 2022-02-24
merged! will be making a new npm version ASAP!