@pyramation 2022-02-24
@darora thank you for your contribution! Please try this new version :) 
```
+ libpg-query@13.2.0
```